Fix PR feedback: re-enable imports, consolidate utilities
- Revert Package.swift to use remote GTFSModel dependency
- Re-enable calendar_dates and directions imports
- Merge DistanceCalculator into StopTimeInterpolator as private method
- Remove duplicate DatabaseHelper, use GTFSModel.DatabaseHelper instead
- Update main.swift to call StopTimeInterpolator directly
