+++ Package.swift
@@ -14,8 +14,7 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-argument-parser", from: "1.0.0"),
         .package(name: "GRDB", url: "https://github.com/groue/GRDB.swift.git", from: "6.29.1"),
         .package(url: "https://github.com/yaslab/CSV.swift.git", .upToNextMinor(from: "2.4.3")),
-        // Use local GTFSModel for development
-        .package(path: "../GTFSModel")
+        .package(url: "https://github.com/jogi/GTFSModel", .branch("main"))
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
