+++ Sources/gtfs-importer/Utilities/StopTimeInterpolator.swift
@@ -11,6 +11,38 @@ import GTFSModel
 import OSLog
 
 enum StopTimeInterpolator {
+    /// Earth radius in meters
+    private static let earthRadius: Double = 6378135
+
+    /// Compute approximate distance between two points in meters using Haversine formula.
+    /// Assumes the Earth is a sphere.
+    ///
+    /// - Parameters:
+    ///   - lat1: Latitude of first point in degrees
+    ///   - lon1: Longitude of first point in degrees
+    ///   - lat2: Latitude of second point in degrees
+    ///   - lon2: Longitude of second point in degrees
+    /// - Returns: Distance in meters
+    private static func approximateDistance(
+        lat1: Double,
+        lon1: Double,
+        lat2: Double,
+        lon2: Double
+    ) -> Double {
+        // Convert degrees to radians
+        let lat1Rad = lat1 * .pi / 180
+        let lon1Rad = lon1 * .pi / 180
+        let lat2Rad = lat2 * .pi / 180
+        let lon2Rad = lon2 * .pi / 180
+
+        // Haversine formula
+        let dlat = sin(0.5 * (lat2Rad - lat1Rad))
+        let dlng = sin(0.5 * (lon2Rad - lon1Rad))
+        let x = dlat * dlat + dlng * dlng * cos(lat1Rad) * cos(lat2Rad)
+
+        return earthRadius * (2 * atan2(sqrt(x), sqrt(max(0.0, 1.0 - x))))
+    }
+
     struct StopTimeRecord: Codable, FetchableRecord {
         var tripId: String
         var arrivalTime: String?
@@ -111,7 +143,7 @@ enum StopTimeInterpolator {
                 // Find the next timepoint and calculate total distance
                 if i + 1 < stopTimes.count {
                     var k = i + 1
-                    distanceBetweenTimepoints += DistanceCalculator.approximateDistance(
+                    distanceBetweenTimepoints += approximateDistance(
                         lat1: stopTimes[k-1].stopLat,
                         lon1: stopTimes[k-1].stopLon,
                         lat2: stopTimes[k].stopLat,
@@ -121,7 +153,7 @@ enum StopTimeInterpolator {
                     while k < stopTimes.count && (stopTimes[k].arrivalTime == nil || stopTimes[k].arrivalTime!.isEmpty) {
                         k += 1
                         if k < stopTimes.count {
-                            distanceBetweenTimepoints += DistanceCalculator.approximateDistance(
+                            distanceBetweenTimepoints += approximateDistance(
                                 lat1: stopTimes[k-1].stopLat,
                                 lon1: stopTimes[k-1].stopLon,
                                 lat2: stopTimes[k].stopLat,
@@ -146,7 +178,7 @@ enum StopTimeInterpolator {
                 }
 
                 // Calculate distance from previous stop
-                distanceTraveledBetweenTimepoints += DistanceCalculator.approximateDistance(
+                distanceTraveledBetweenTimepoints += approximateDistance(
                     lat1: stopTimes[i-1].stopLat,
                     lon1: stopTimes[i-1].stopLon,
                     lat2: st.stopLat,
